Crate snowbridge_pallet_ethereum_client

Ethereum Beacon Client

A light client that verifies consensus updates signed by the sync committee of the beacon chain.

Extrinsics

Governance

Call::force_checkpoint: Set the initial trusted consensus checkpoint.

Call::set_operating_mode: Set the operating mode of the pallet. Can be used to disable processing of consensus updates.

Consensus Updates

Call::submit: Submit a finalized beacon header with an optional sync committee update

Re-exports

pub use weights::WeightInfo;

pub use config::SLOTS_PER_HISTORICAL_ROOT;

pub use pallet::*;

Modules

config

functions

impls

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

types

weights

Autogenerated weights for ethereum_beacon_client

Constants

LOG_TARGET