Module snowbridge_pallet_ethereum_client::config

Constants

BLOCK_ROOTS_DEPTH

BLOCK_ROOTS_INDEX

Generalized Indices related to Merkle proofs get_generalized_index(BeaconState, ‘block_roots’)

BLOCK_ROOTS_SUBTREE_INDEX

BLOCK_ROOT_AT_INDEX_DEPTH

The index of the block_roots field in the beacon state tree.

CURRENT_SYNC_COMMITTEE_DEPTH

CURRENT_SYNC_COMMITTEE_INDEX

get_generalized_index(BeaconState, ‘current_sync_committee’)

CURRENT_SYNC_COMMITTEE_SUBTREE_INDEX

DOMAIN_SYNC_COMMITTEE

DomainType(‘0x07000000’) https://github.com/ethereum/consensus-specs/blob/dev/specs/altair/beacon-chain.md#domain-types

EPOCHS_PER_SYNC_COMMITTEE_PERIOD

256 epochs in a sync committee period. Frequency of sync committee (subset of Ethereum validators) change is every ~27 hours.

EXECUTION_HEADER_DEPTH

EXECUTION_HEADER_INDEX

get_generalized_index(BeaconBlockBody, ‘execution_payload’)

EXECUTION_HEADER_SUBTREE_INDEX

FINALIZED_ROOT_DEPTH

FINALIZED_ROOT_INDEX

get_generalized_index(BeaconState, ‘finalized_checkpoint’, ‘root’)

FINALIZED_ROOT_SUBTREE_INDEX

MAX_BRANCH_PROOF_SIZE

Sanity value to constrain the max size of a merkle branch proof.

MAX_EXTRA_DATA_BYTES

Sizes related to SSZ encoding

MAX_FEE_RECIPIENT_SIZE

MAX_LOGS_BLOOM_SIZE

NEXT_SYNC_COMMITTEE_DEPTH

NEXT_SYNC_COMMITTEE_INDEX

get_generalized_index(BeaconState, ‘next_sync_committee’)

NEXT_SYNC_COMMITTEE_SUBTREE_INDEX

PUBKEY_SIZE

Validators public keys are 48 bytes.

SIGNATURE_SIZE

Signatures produced by validators are 96 bytes.

SLOTS_PER_EPOCH

Defined in https://github.com/ethereum/consensus-specs/tree/f1dff5f6768608d890fc0b347e548297fc3e1f1c/presets/mainnet There are 32 slots in an epoch. An epoch is 6.4 minutes long.

SLOTS_PER_HISTORICAL_ROOT

The size of the block root array in the beacon state, used for ancestry proofs.

SYNC_COMMITTEE_BITS_SIZE

An array of sync committee block votes, one bit representing the vote of one validator.

SYNC_COMMITTEE_SIZE

A sync committee contains 512 randomly selected validators.