snowbridge_pallet_ethereum_client

Module mock

Structs

ChainForkVersions

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

RuntimeGenesisConfig

RuntimeOrigin

The runtime origin type representing the origin of a call.

Test

Enums

OriginCaller

RuntimeCall

The aggregated runtime call type.

RuntimeError

RuntimeEvent

RuntimeFreezeReason

A reason for placing a freeze on funds.

RuntimeHoldReason

A reason for placing a hold on funds.

RuntimeLockId

An identifier for each lock placed on funds.

RuntimeSlashReason

A reason for slashing funds.

RuntimeTask

An aggregation of all Task enums across all pallets included in the current runtime.

RuntimeViewFunction

Runtime query type.

Constants

FREE_SLOTS_INTERVAL

Functions

get_message_verification_payload

initialize_storage

load_checkpoint_update_fixture

load_execution_proof_fixture

load_finalized_header_update_fixture

load_next_finalized_header_update_fixture

load_next_sync_committee_update_fixture

load_sync_committee_update_fixture

load_sync_committee_update_period_0

load_sync_committee_update_period_0_newer_fixture

load_sync_committee_update_period_0_older_fixture

new_tester

Type Aliases

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

EthereumBeaconClient

System

Timestamp