Module snowbridge_pallet_ethereum_client::pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

The main components of this pallet are:

Pallet, which implements all of the dispatchable extrinsics of the pallet, among other public functions.

The subset of the functions that are dispatchable can be identified either in the dispatchables module or in the Call enum.

storage_types, which contains the list of all types that are representing a storage item. Otherwise, all storage items are listed among Type Definitions.

Config, which contains the configuration trait of this pallet.

Event and Error, which are listed among the Enums.

Re-exports

pub use __tt_error_token_1 as tt_error_token;

pub use __tt_default_parts_7 as tt_default_parts;

pub use __tt_extra_parts_7 as tt_extra_parts;

pub use __tt_default_parts_v2_7 as tt_default_parts_v2;

Modules

dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

Structs

MaxFinalizedHeadersToKeep

Pallet

The Pallet struct, the main type that implements traits and standalone functions within the pallet.

Enums

Call

Contains a variant per dispatchable extrinsic that this pallet has.

Error

The Error enum of this pallet.

Event

The Event enum of this pallet

Traits

Config

Configuration trait of this pallet.

Type Aliases

CurrentSyncCommittee

Sync committee for current period

FinalizedBeaconState

Beacon state by finalized block root

FinalizedBeaconStateIndex

Finalized Headers: Current position in ring buffer

FinalizedBeaconStateMapping

Finalized Headers: Mapping of ring buffer index to a pruning candidate

InitialCheckpointRoot

Latest imported checkpoint root

LatestFinalizedBlockRoot

Latest imported finalized block root

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.

NextSyncCommittee

Sync committee for next period

OperatingMode

The current operating mode of the pallet.

ValidatorsRoot

Storage type is StorageValue with value type H256.