Module snowbridge_pallet_ethereum_client::pallet::storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes.

To see the actual storage type, find a struct with the same name at the root of the pallet, in the list of Type Definitions.

Structs

CurrentSyncCommittee

Sync committee for current period

FinalizedBeaconState

Beacon state by finalized block root

FinalizedBeaconStateIndex

Finalized Headers: Current position in ring buffer

FinalizedBeaconStateMapping

Finalized Headers: Mapping of ring buffer index to a pruning candidate

InitialCheckpointRoot

Latest imported checkpoint root

LatestFinalizedBlockRoot

Latest imported finalized block root

NextSyncCommittee

Sync committee for next period

OperatingMode

The current operating mode of the pallet.

ValidatorsRoot

Warning: Doc-Only