Type Alias snowbridge_pallet_ethereum_client::types::Update

pub type Update = Update<SC_SIZE, SC_BITS_SIZE>;

Aliased Type

struct Update {
    pub attested_header: BeaconHeader,
    pub sync_aggregate: SyncAggregate<512, 64>,
    pub signature_slot: u64,
    pub next_sync_committee_update: Option<NextSyncCommitteeUpdate<512>>,
    pub finalized_header: BeaconHeader,
    pub finality_branch: Vec<H256>,
    pub block_roots_root: H256,
    pub block_roots_branch: Vec<H256>,
}

Fields
attested_header: BeaconHeader

A recent header attesting to the finalized header, using its state_root.

sync_aggregate: SyncAggregate<512, 64>

The signing data that the sync committee produced for this attested header, including who participated in the vote and the resulting signature.

signature_slot: u64

The slot at which the sync aggregate can be found, typically attested_header.slot + 1, if the next slot block was not missed.

next_sync_committee_update: Option<NextSyncCommitteeUpdate<512>>

The next sync committee for the next sync committee period, if present.

finalized_header: BeaconHeader

The latest finalized header.

finality_branch: Vec<H256>

The merkle proof testifying to the finalized header, using the attested_header.state_root as tree root.

block_roots_root: H256

The finalized_header’s block_roots root in the beacon state, used for ancestry proofs.

block_roots_branch: Vec<H256>

The merkle path to prove the block_roots_root value.